Remove unnecessary methods from sample service for E2E test.
